+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Remove `CircleCI::CoverageReporter::Client#configuration`
 
 ## [0.2.0] - 2017-03-14
 ### Added